+++ src/app/task.h
@@ -7,13 +7,13 @@ class Task{
 private:
 public:
 const int task_id_;
-int task_time_;
+double task_time_;
 double priority_;
 string title_;
 static int global_id;
 int days_left_;
 
-Task(string title, int task_time, int days_left);
+Task(string title, double task_time, int days_left);
 
 };
